Inline query-async
