--- packages/reactive-element/src/std-decorators/query-async.ts
@@ -1,63 +1,0 @@
-/**
- * @license
- * Copyright 2017 Google LLC
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-/*
- * IMPORTANT: For compatibility with tsickle and the Closure JS compiler, all
- * property decorators (but not class decorators) in this file that have
- * an @ExportDecoratedItems annotation must be defined as a regular function,
- * not an arrow function.
- */
-
-import type {ReactiveElement} from '../reactive-element.js';
-import type {Interface} from '../decorators/base.js';
-
-/**
- * A property decorator that converts a class property into a getter that
- * returns a promise that resolves to the result of a querySelector on the
- * element's renderRoot done after the element's `updateComplete` promise
- * resolves. When the queried property may change with element state, this
- * decorator can be used instead of requiring users to await the
- * `updateComplete` before accessing the property.
- *
- * @param selector A DOMString containing one or more selectors to match.
- *
- * See: https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelector
- *
- * ```ts
- * class MyElement {
- *   @queryAsync('#first')
- *   first: Promise<HTMLDivElement>;
- *
- *   render() {
- *     return html`
- *       <div id="first"></div>
- *       <div id="second"></div>
- *     `;
- *   }
- * }
- *
- * // external usage
- * async doSomethingWithFirst() {
- *  (await aMyElement.first).doSomething();
- * }
- * ```
- * @category Decorator
- */
-// TODO(justinfagnani): infer a more precise return type when the query is
-// a tagname.
-export const queryAsync =
-  (selector: string) =>
-  <C extends Interface<ReactiveElement>, V extends Promise<Element | null>>(
-    _target: ClassAccessorDecoratorTarget<C, V>,
-    _context: ClassAccessorDecoratorContext<C, V>
-  ) => {
-    return {
-      async get(this: C) {
-        await this.updateComplete;
-        return this.renderRoot?.querySelector(selector) ?? null;
-      },
-    };
-  };
